+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Show password requirements in relevant screens
 * Added Information about Instance to Login Screen. Can be configured via dockerapp-Config.
 * Moved Email Functionality to API, removed RailMail from Project
 * Changed Username, displayname spec to allow common chars (-., umlauts)